refactor: componentsフォルダの構成を整理
