+++ src/components/layout/Layout.jsx
@@ -1,4 +1,4 @@
-import Header from '../Header';
+import Header from './Header';
 import { Outlet } from 'react-router-dom';
 
 export default function Layout() {
